+++ frontend/src/pages/Epics/List/columns.jsx
@@ -167,7 +167,6 @@ export const columnsEpic = () => {
 		[epics],
 	);
 	// Add actions column for Superadmin
-	// TODO: View projects under epic
 	const columnsWithActions = useMemo(() => {
 		if (user?.data?.role === "Superadmin" || user?.data?.role === "Admin" || user?.data?.role === "Manager") {
 			return [